+++ app/src/main/java/fr/machada/bpm/pro/event/OnFBShareFCEvent.java
@@ -1,0 +1,16 @@
+package fr.machada.bpm.pro.event;
+
+/**
+ * Created by macha on 11/03/16.
+ */
+public class OnFBShareFCEvent {
+    public final int value;
+
+    public OnFBShareFCEvent(int value) {
+        this.value = value;
+    }
+
+    public int getValue() {
+        return value;
+    }
+}
